Add: Seeder Factory & Models

+++ database/migrations/2023_12_25_012718_create_conveyors_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('nama');
             $table->string('tipe');
             $table->string('status');
-            $table->integer('kapasitas');
+            $table->float('kapasitas');
             $table->timestamps();
         });
     }
